+++ src/router/AppRouter.jsx
@@ -35,6 +35,8 @@ const PurchaseDetail = lazy(() => import('../pages/purchases/PurchaseDetail'));
 const TransferList = lazy(() => import('../pages/transfers/TransferList'));
 const TransferForm = lazy(() => import('../pages/transfers/TransferForm'));
 const TransferDetail = lazy(() => import('../pages/transfers/TransferDetail'));
+const CustomerList = lazy(() => import('../pages/customers/CustomerList'));
+const CustomerDetail = lazy(() => import('../pages/customers/CustomerDetail'));
 const NotFound404 = lazy(() => import('../pages/errors/NotFound404'));
 
 function RouteFallback() {
@@ -84,6 +86,8 @@ function AppRouter() {
               <Route path="/transfers" element={<TransferList />} />
               <Route path="/transfers/new" element={<TransferForm />} />
               <Route path="/transfers/:id" element={<TransferDetail />} />
+              <Route path="/customers" element={<CustomerList />} />
+              <Route path="/customers/:id" element={<CustomerDetail />} />
             </Route>
           </Route>